backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
opencv-python==4.8.1.78
moviepy==1.0.3
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0

